requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blockbuster==1.5.25
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
cryptography==44.0.3
forbiddenfruit==0.1.4
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-tools==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain-core==0.3.79
langchain-ollama==0.3.10
langgraph==1.0.1
langgraph-api==0.5.16
langgraph-checkpoint==3.0.1
langgraph-cli==0.4.7
langgraph-prebuilt==1.0.1
langgraph-runtime-inmem==0.18.0
langgraph-sdk==0.2.9
langgraph-supervisor==0.0.30
langgraph-swarm==0.0.15
langsmith==0.4.43
ollama==0.6.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pillow==12.0.0
protobuf==6.33.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
sse-starlette==2.1.3
starlette==0.50.0
structlog==25.5.0
tenacity==9.1.2
truststore==0.10.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
